This function retrieves time series data.

Given the station identification number(s), this function retrieves data (time series in zoo format with accompanying metadata) from the WaterML2 service on the NRFA database. The time series can be of two types: cmr (catchment mean rainfall, monthly) or gdf (gauged daily flows, daily).

get_ts(
  id,
  type,
  metadata = FALSE,
  cl = NULL,
  verbose = FALSE,
  full_info = FALSE
)

Arguments

id
station identification number(s), each number should be in the range [3002,236051].

type
The following data-types are available:

gdf = Gauged daily flows

gmf = Gauged monthly flows

ndf = Naturalised daily flows

nmf = Naturalised monthly flows

cdr = Catchment daily rainfall

cdr-d = Catchment daily rainfall distance to rain gauge

cmr = Catchment monthly rainfall

pot-stage = Peaks over threshold stage

pot-flow = Peaks over threshold flow

gauging-stage = Gauging stage

gauging-flow = Gauging flow

amax-stage = Annual maxima stage

amax-flow = Annual maxima flow

metadata
Logical, FALSE by default. If metadata = TRUE means that the result for a single station is a list with two elements: data (the time series) and meta (metadata).

cl
(optional) This is a cluster object, created by the parallel package. This is set to NULL by default, which sends sequential calls to the server.

verbose
(FALSE by default). If set to TRUE prints GET request on the console.

full_info
Logical, FALSE by default. If full_info = TRUE, the function will retrieve info on rejected periods and return a data frame rather than a time series.

Value

list composed of as many objects as in the list of station identification numbers. Each object can be accessed using their names or indexes (e.g. x[[1]], x[[2]], and so forth). Each object contains a time series of class zoo/xts.

Author

Claudia Vitolo

Examples

if (FALSE) {
  get_ts(18019, type = "cmr")
  get_ts(c(54022,54090,54091), type = "cmr")
  get_ts(18019, type = "gdf")
  get_ts(c(54022,54090,54091), type = "gdf")
  plot(get_ts(id = 23001, type = "ndf"))
  plot(get_ts(id = 23001, type = "nmf"))
}