Plot rainfall and flow for a given station

This function retrieves rainfall and flow time series for a given catchment, divides the flow by the catchment area and converts it to mm/day to that it can be comparable with the rainfall (mm/month). Finally it generates a plots combining rainfall and flow information.

plot_rain_flow(id = NULL, rain = NULL, flow = NULL, area = NULL, title = "")

Arguments

id
Station identification number

rain
Rainfall time series, measured in mm/month

flow
Flow time series, measured in m3/s

area
Catchment area in Km2

title
(optional) Plot title

Value

Plot rainfall and flow for a given station

Examples

if (FALSE) {
  plot_rain_flow(id = 54090)
}